Move Money non-static behavior into new namespace
Move Money non-static class into GameServer/Finance/Money.cs
Move Currency.cs into Finance folder as well
