+++ GameServer/Finance/Currency.cs
