+++ GameServer/gameutils/Money.cs
@@ -17,7 +17,6 @@
  *
  */
 using System.Text;
-using DOL.GS.Finance;
 using DOL.GS.ServerProperties;
 using DOL.Language;
 
@@ -25,27 +24,6 @@ namespace DOL.GS
 {
 	public class Money
 	{
-        public long Amount { get; private set; }
-        public Currency Currency { get; private set; }
-
-        private Money() { }
-
-        public static Money Mint(long value, Currency type)
-            => new Money() { Amount = value, Currency = type };
-
-        public override bool Equals(object obj)
-        {
-            if (obj is Money otherCurrency)
-            {
-                var areOfSameValue = otherCurrency.Amount == this.Amount;
-                var areOfSameType = otherCurrency.Currency.Equals(Currency);
-                return areOfSameType && areOfSameValue;
-            }
-            return false;
-        }
-
-        public override int GetHashCode() => base.GetHashCode();
-
 		public static int GetMithril(long money)
 		{
 			return (int)(money / 100L / 100L / 1000L / 1000L % 1000L);
